test(api): bundle supabase browser lib
Bundling the `@supabase/auth-helpers-shared` library to use in the
browser context allows us to initialize and save the authorization
cookies for subsequent tests.

Since our application will be using the auth helpers on both the server
and client, the auth token and refresh token will be sent over in the
"supabase-auth-token" cookie.